test(docgen): consolidate class property example coverage
