+++ packages/tools/docgen/test/Core.test.ts
@@ -1,8 +1,15 @@
+import * as Configuration from "@effect/docgen/Configuration"
 import * as Core from "@effect/docgen/Core"
+import * as Domain from "@effect/docgen/Domain"
 import * as Parser from "@effect/docgen/Parser"
 import * as NodeServices from "@effect/platform-node/NodeServices"
 import { assert, describe, it } from "@effect/vitest"
 import * as Effect from "effect/Effect"
+import * as FileSystem from "effect/FileSystem"
+import * as Path from "effect/Path"
+import * as Sink from "effect/Sink"
+import * as Stream from "effect/Stream"
+import * as ChildProcessSpawner from "effect/unstable/process/ChildProcessSpawner"
 import * as ast from "ts-morph"
 
 const assertFencedCode = (
@@ -13,7 +20,139 @@ const assertFencedCode = (
   assert.deepStrictEqual(Core.extractFencedCode(markdown), [expectedExamples, expectedWarnings])
 }
 
+const assertExampleFiles = (source: string, expected: ReadonlyArray<string>, runExamples = true) =>
+  Effect.gen(function*() {
+    const fs = yield* FileSystem.FileSystem
+    const path = yield* Path.Path
+    const cwd = yield* fs.makeTempDirectoryScoped()
+    const config: Configuration.ConfigurationShape = {
+      projectName: "docgen",
+      projectHomepage: "https://example.test",
+      srcLink: "https://example.test",
+      srcDir: path.relative(process.cwd(), path.join(cwd, "src")),
+      outDir: path.relative(process.cwd(), path.join(cwd, "docs")),
+      theme: Configuration.DEFAULT_THEME,
+      enableSearch: true,
+      enforceDescriptions: false,
+      enforceExamples: false,
+      enforceVersion: false,
+      tscExecutable: "tsc",
+      runExamples,
+      exclude: [],
+      parseCompilerOptions: {},
+      examplesCompilerOptions: Configuration.defaultCompilerOptions
+    }
+    yield* fs.makeDirectory(config.srcDir)
+    yield* fs.writeFileString(path.join(config.srcDir, "Box.ts"), source)
+
+    const commands: Array<string> = []
+    // Inspect the generated inputs without invoking a compiler or executing examples.
+    const spawner = ChildProcessSpawner.make((command) =>
+      Effect.gen(function*() {
+        if (command._tag !== "StandardCommand") return yield* Effect.die("Unexpected pipeline")
+        commands.push(command.command)
+        const directory = path.join(config.outDir, "examples")
+        const names = (yield* fs.readDirectory(directory))
+          .filter((name) => name.endsWith(".ts") && name !== "index.ts").sort()
+        const contents = yield* Effect.forEach(names, (name) => fs.readFileString(path.join(directory, name)))
+        assert.deepStrictEqual(contents.sort(), [...expected].sort())
+        const index = yield* fs.readFileString(path.join(directory, "index.ts"))
+        assert.deepStrictEqual(index.trim().split("\n").sort(), names.map((name) => `import './${name.slice(0, -3)}'`))
+        return ChildProcessSpawner.makeHandle({
+          pid: ChildProcessSpawner.ProcessId(1),
+          exitCode: Effect.succeed(ChildProcessSpawner.ExitCode(0)),
+          isRunning: Effect.succeed(false),
+          kill: () => Effect.void,
+          stdin: Sink.drain,
+          stdout: Stream.empty,
+          stderr: Stream.empty,
+          all: Stream.empty,
+          getInputFd: () => Sink.drain,
+          getOutputFd: () => Stream.empty,
+          unref: Effect.succeed(Effect.void)
+        })
+      })
+    )
+    yield* Core.program.pipe(
+      Effect.provideService(Configuration.Configuration, config),
+      Effect.provideService(Domain.Process, {
+        cwd: Effect.succeed(process.cwd()),
+        platform: Effect.succeed(process.platform),
+        argv: Effect.succeed([]),
+        env: Effect.succeed({})
+      }),
+      Effect.provideService(ChildProcessSpawner.ChildProcessSpawner, spawner)
+    )
+    const executables = expected.length === 0 ? [] : runExamples ? ["tsc", "tsx"] : ["tsc"]
+    assert.deepStrictEqual(commands, executables.map((name) => process.platform === "win32" ? `${name}.cmd` : name))
+  }).pipe(Effect.provide(NodeServices.layer))
+
 describe("Core", () => {
+  describe("class property examples", () => {
+    for (const runExamples of [true, false]) {
+      it.effect(`collects property-only examples with runExamples: ${runExamples}`, () =>
+        assertExampleFiles(
+          `export class Box {
+            /**
+             * @example
+             * ~~~ts
+             * const value = 1
+             * ~~~
+             */
+            readonly value = 1
+          }`,
+          ["const value = 1"],
+          runExamples
+        ))
+    }
+
+    it.effect("collects property descriptions and example tags alongside methods, excluding skipped fences", () =>
+      assertExampleFiles(
+        `export class Box {
+          /**
+           * ~~~ts
+           * const description = 1
+           * ~~~
+           * @example
+           * ~~~ts
+           * const first = 1
+           * ~~~
+           * @example
+           * ~~~ts
+           * const second = 2
+           * ~~~
+           * @example
+           * ~~~ts skip-type-checking
+           * const skipped: string = 1
+           * ~~~
+           */
+          readonly value = 1
+          /**
+           * @example
+           * ~~~ts
+           * const method = 3
+           * ~~~
+           */
+          read() { return this.value }
+        }`,
+        ["const description = 1", "const first = 1", "const second = 2", "const method = 3"]
+      ))
+
+    it.effect("does not request validation when all property examples are skipped", () =>
+      assertExampleFiles(
+        `export class Box {
+          /**
+           * @example
+           * ~~~ts skip-type-checking
+           * const skipped: string = 1
+           * ~~~
+           */
+          readonly value = 1
+        }`,
+        []
+      ))
+  })
+
   it.effect("preserves a type alias name and declared type parameters", () => {
     const sourceFile = new ast.Project({ useInMemoryFileSystem: true }).createSourceFile(
       "Identity.ts",
